+++ routes/api/produits.php
@@ -25,23 +25,23 @@ use App\Http\Controllers\Produit\ProduitArchivedListController;
 */
 
 Route::prefix('produits')->group(function () {
-    // Liste et consultation
-    Route::get('/', ProduitIndexController::class);
-    Route::get('/archived', ProduitArchivedListController::class);
-    Route::get('/search', ProduitSearchController::class);
-    Route::get('/statistics', ProduitStatisticsController::class);
-    Route::get('/{id}', ProduitShowController::class)->where('id', '[0-9]+');
+    // Lecture
+    Route::get('/', ProduitIndexController::class)->middleware('permission:produits.read');
+    Route::get('/archived', ProduitArchivedListController::class)->middleware('permission:produits.read');
+    Route::get('/search', ProduitSearchController::class)->middleware('permission:produits.read');
+    Route::get('/statistics', ProduitStatisticsController::class)->middleware('permission:produits.read');
+    Route::get('/{id}', ProduitShowController::class)->where('id', '[0-9]+')->middleware('permission:produits.read');
 
-    // CRUD
-    Route::post('/', ProduitStoreController::class);
-    Route::put('/{id}', ProduitUpdateController::class)->where('id', '[0-9]+');
-    Route::delete('/{id}', ProduitDestroyController::class)->where('id', '[0-9]+');
+    // Création
+    Route::post('/', ProduitStoreController::class)->middleware('permission:produits.create');
 
-    // Actions sur le stock
-    Route::patch('/{id}/stock', ProduitUpdateStockController::class)->where('id', '[0-9]+');
+    // Mise à jour
+    Route::put('/{id}', ProduitUpdateController::class)->where('id', '[0-9]+')->middleware('permission:produits.update');
+    Route::patch('/{id}/stock', ProduitUpdateStockController::class)->where('id', '[0-9]+')->middleware('permission:produits.update');
+    Route::patch('/{id}/status', ProduitChangeStatusController::class)->where('id', '[0-9]+')->middleware('permission:produits.update');
+    Route::patch('/{id}/archive', ProduitArchiveController::class)->where('id', '[0-9]+')->middleware('permission:produits.update');
+    Route::patch('/{id}/unarchive', ProduitUnarchiveController::class)->where('id', '[0-9]+')->middleware('permission:produits.update');
 
-    // Actions sur le statut
-    Route::patch('/{id}/status', ProduitChangeStatusController::class)->where('id', '[0-9]+');
-    Route::patch('/{id}/archive', ProduitArchiveController::class)->where('id', '[0-9]+');
-    Route::patch('/{id}/unarchive', ProduitUnarchiveController::class)->where('id', '[0-9]+');
+    // Suppression
+    Route::delete('/{id}', ProduitDestroyController::class)->where('id', '[0-9]+')->middleware('permission:produits.delete');
 });
